+++ NEWS.md
@@ -6,6 +6,7 @@
   values with `argN` anymore
   ([#336](https://github.com/company-mode/company-mode/issues/336)). This
   affects `company-clang` and all third-party backends that use this function.
+* Likewise for `company-clang-objc-templatify`.
 * `company-template-add-field` calling convention has changed.
 * New user option `company-dabbrev-ignore-invisible`.
 * `company-ropemacs` was removed. `ropemacs` supports completion via